Added new features to power supply
Support ADC to read test voltage values. You can now choose to show PS voltage and current or test voltage values on each line. Currently this is done by commenting and uncommenting lines. But eventually it will be  controlled by the Buttons and a menu. ADC takes samples from each of 4 channels and accumulates 16 samples before taking the average and storing that.

Also added support Buttons. But for now these are not hooked up to anything. Documented all the port bits for the circuit. Modularized updates and display logic.